requirements.txt
os
requests
flask
io
pillow
google.generativeai